+++ src/RavenNest.DataModels/Village.cs
@@ -7,7 +7,7 @@ namespace RavenNest.DataModels
         private Guid userId; public Guid UserId { get => userId; set => Set(ref userId, value); }
         private string name; public string Name { get => name; set => Set(ref name, value); }
         private int level; public int Level { get => level; set => Set(ref level, value); }
-        private long experience; public long Experience { get => experience; set => Set(ref experience, value); }
+        private double experience; public double Experience { get => experience; set => Set(ref experience, value); }
         private Guid resourcesId; public Guid ResourcesId { get => resourcesId; set => Set(ref resourcesId, value); }
     }
 }
